SFC/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
mysqlclient==2.2.4
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
sentencepiece==0.2.0
setuptools==69.5.1
sqlparse==0.4.4
uritemplate==4.1.1
wheel==0.43.0
